Separated plugin implementations from base classes
Removed unused PCL dependency

Added orientation to pose on path planner

Moved concave hull and ear clipping to source and renamed references.

Cleaned up autogenerated code and included variable names in prototype

Made mesh_importer a unique pointer and changed from a private class member to a local object
