+++ meshing_plugins_base/include/meshing_plugins_base/meshing_base.h
@@ -1,34 +1,28 @@
-#ifndef MESHING_BASE_H_
-#define MESHING_BASE_H_
+#ifndef MESHING_PLUGINS_BASE_H_
+#define MESHING_PLUGINS_BASE_H_
 
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl/PolygonMesh.h>
 
-namespace meshing_base
+namespace meshing_plugins_base
 {
-  typedef pcl::PointXYZRGB Point;
-  typedef pcl::PointCloud<Point> PointCloud;
-
   class MeshingBase
   {
   public:
     /**
      * @brief Initialize the mesher
-     * @details [long description]
-     * 
      * @param  pcl::PointCloud input cloud for the mesher
      */
-    virtual void init(PointCloud) = 0;
+    virtual void init(pcl::PointCloud<pcl::PointXYZRGB> input) = 0;
 
     /**
      * @brief Generates a pcl::PolygonMesh from the input PointCloud
-     * @details [long description]
-     * @param  [description]
+     * @param  mesh: destination variable for the resulting mesh
      * @return true if the generation was successful, false if it failed
      */
-    virtual bool generateMesh(pcl::PolygonMesh&) = 0;
+    virtual bool generateMesh(pcl::PolygonMesh& mesh) = 0;
   };
-} // end mesher_base
+} // end namespace meshing_plugins_base
 
-#endif
+#endif // end MESHING_PLUGINS_BASE_H_
