+++ meshing_plugins_base/package.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<package>
+  <name>meshing_plugins_base</name>
+  <version>0.1.0</version>
+  <description>
+    Base class for plugins which convert pcl::PointCloud to pcl::PolygonMesh
+  </description>
+
+  <maintainer email="awood@swri.org">Aaron Wood</maintainer>
+  <license>Apache 2.0</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <build_depend>pcl_ros</build_depend>
+  <build_depend>roscpp</build_depend>
+
+  <run_depend>pcl_ros</run_depend>
+  <run_depend>roscpp</run_depend>
+</package>
