+++ path_planning_plugins/package.xml
@@ -16,6 +16,9 @@
   <build_depend>geometry_msgs</build_depend>
   <build_depend>godel_msgs</build_depend>
   <build_depend>godel_process_path_generation</build_depend>
+  <build_depend>path_planning_plugins_base</build_depend>
+  <build_depend>pcl_ros</build_depend>
+  <build_depend>pluginlib</build_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>tf</build_depend>
 
@@ -24,10 +27,13 @@
   <run_depend>geometry_msgs</run_depend>
   <run_depend>godel_msgs</run_depend>
   <run_depend>godel_process_path_generation</run_depend>
+  <run_depend>path_planning_plugins_base</run_depend>
+  <run_depend>pcl_ros</run_depend>
+  <run_depend>pluginlib</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>tf</run_depend>
 
   <export>
-    <path_planning_plugins plugin="${prefix}/plugin.xml"/>
+    <path_planning_plugins_base plugin="${prefix}/plugin.xml"/>
   </export>
 </package>
