+++ path_planning_plugins_base/package.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<package>
+  <name>path_planning_plugins_base</name>
+  <version>0.0.1</version>
+  <description>
+    Base class for meshing plugins
+  </description>
+
+  <maintainer email="awood@swri.org">Aaron Wood</maintainer>
+  <license>Apache 2.0</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <build_depend>geometry_msgs</build_depend>
+  <build_depend>pcl_ros</build_depend>
+  <build_depend>roscpp</build_depend>
+
+  <run_depend>geometry_msgs</run_depend>
+  <run_depend>pcl_ros</run_depend>
+  <run_depend>roscpp</run_depend>
+</package>
